slack-clone/Backend Django/slack_clone/requirements.txt
djangorestframework-simplejwt
djangorestframework
channels 
channels
django-cors-headers
psycopg2
django 
djangorestframework 
channels 
djangorestframework-simplejwt 
psycopg2-binary 
channels-redis 
django-cors-headers 
psycopg2
django 
djangorestframework 
channels
psycopg2-binary
